Language changes & validation
Added support for initializers for variables.
Changed CompoundResource to CompoundVariable, making it inherit most of it's guts from Variable.
Began with SPIRV validation, but right now I am considering abstracting most of it too because there is a lot of shared boilerplate code there, and very little which pertains specifically to SPIRV. In fact, for the validation step, I will look into generalizing that whole process into just a single validator. The only thing that strikes me is the vulkan binding model (descriptor set and binding), available image formats and qualifiers etc. But these 'allowed attributes' could be provided by a language specific validator while the actual validation code is generic.
All symbols have a __Resolved state which is the product of validation, basically when we unravel all expressions, setup symbol-internal states and enums, and prepare the AST for code and binary generation.