Major AnyFX changes
All objects are handled as symbols, and are typechecked and output in the order they appear in the original shader code. This helps with error location as well as enforces ordering rules, allowing for loser ordering rules when compiling.
Symbols are allocated when parsed, and never copied around, which could improve performance a bit. To get truly better performance, we should use an arena allocator when creating symbols while parsing. '
Moved all string parsing from the grammar to the code. This removes the need to change the grammar when adding qualifiers or types, which is a big win.
Added support for buffer pointers, as supported by GL_EXT_buffer_reference. This allows for the programmer to declare a struct as a reference to a buffer, which when used as a member in an rw_buffer will not expand it to a structure type, but will be a buffer_reference in GLSL and a uint64 in C.