Smarter binding assignment
Bindings used to simply increment a counter, but with this new system a 64 bit uint is used to flag for free slots. Whenever a new slot is consumed, a free bit is searched for using ctz and then the bit is used. This allows for aliasing high slots without pushing all future bindings continue from there.